# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Ascension Quest API
  version: 0.1.0
  description: "Ascension Quest: Legends of Aethoria API"
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /player:
    get:
      operationId: getPlayer
      tags: [player]
      summary: Get player profile
      responses:
        "200":
          description: Player profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
    patch:
      operationId: updatePlayer
      tags: [player]
      summary: Update player profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlayerUpdate"
      responses:
        "200":
          description: Updated player
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
  /player/setup:
    post:
      operationId: setupPlayer
      tags: [player]
      summary: Initial character setup — sets name, stat bonuses, and unlocks equipment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlayerSetupInput"
      responses:
        "200":
          description: Player after setup
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
  /player/allocate-stats:
    post:
      operationId: allocateStats
      tags: [player]
      summary: Allocate free stat points
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StatAllocation"
      responses:
        "200":
          description: Updated player stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
...
  /player/respec:
    post:
      operationId: respecPlayer
      tags: [player]
      summary: Reset stat allocation for gold or Respec Scroll
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                method:
                  type: string
                  enum: [gold, scroll]
              required: [method]
      responses:
        "200":
          description: Updated player after respec
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  player:
                    $ref: "#/components/schemas/Player"
                  pointsReturned:
                    type: integer
...
  /player/titles/{id}/equip:
    post:
      operationId: equipTitle
      tags: [player]
      summary: Equip a title
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated player
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
...
  /player/daily-reward:
    get:
      operationId: getDailyReward
      tags: [player]
      summary: Get daily login reward status
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Daily reward status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyRewardStatus"
  /player/daily-reward/claim:
    post:
      operationId: claimDailyReward
      tags: [player]
      summary: Claim today's daily login reward
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Reward claimed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyRewardClaimResult"
...
components:
  schemas:
...
    Player:
      type: object
      required: [id, name, level, rank, xp, xpToNextLevel, hp, maxHp, mp, maxMp, gold, freeStatPoints, stats, activeTitle, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        level:
          type: integer
        rank:
          type: string
          enum: [E, D, C, B, A, S, "National-Level"]
        xp:
          type: integer
        xpToNextLevel:
          type: integer
        hp:
          type: integer
        maxHp:
          type: integer
        mp:
          type: integer
        maxMp:
          type: integer
        gold:
          type: integer
        freeStatPoints:
          type: integer
        streakDays:
          type: integer
        stats:
          $ref: "#/components/schemas/PlayerStats"
        activeTitle:
          type: ["string", "null"]
        penaltyQuestActive:
          type: boolean
        baseClass:
          type: ["string", "null"]
        setupCompleted:
          type: boolean
        createdAt:
          type: string
    PlayerStats:
      type: object
      required: [strength, agility, stamina, vitality, discipline, sense]
      properties:
        strength:
          type: integer
        agility:
          type: integer
        stamina:
          type: integer
        vitality:
          type: integer
        discipline:
          type: integer
        sense:
          type: integer
    PlayerSetupInput:
      type: object
      required: [name, statBonuses, equipmentIds]
      properties:
        name:
          type: string
        statBonuses:
          $ref: "#/components/schemas/PlayerStats"
        equipmentIds:
          type: array
          items:
            type: integer
        baseClass:
          type: string
        goal:
          type: string
        systemScan:
          type: object
          additionalProperties: false
          properties:
            ageYears:
              type: integer
            sex:
              type: string
              enum: [male, female, other]
            heightCm:
              type: number
              nullable: true
            weightKg:
              type: number
              nullable: true
            activityLevel:
              type: string
              enum: [sedentary, light, moderate, active, very_active]
            weightGoal:
              type: string
              enum: [lose, maintain, gain]
            equipmentTypes:
              type: array
              items:
                type: string
    PlayerUpdate:
      type: object
      properties:
        name:
          type: string
    StatAllocation:
      type: object
      required: [allocations]
      properties:
        allocations:
          $ref: "#/components/schemas/PlayerStats"
...
    DailyRewardEntry:
      type: object
      properties:
        day:
          type: integer
        type:
          type: string
        amount:
          type: integer
        label:
          type: string
    DailyRewardCalendarDay:
      type: object
      properties:
        day:
          type: integer
        reward:
          $ref: "#/components/schemas/DailyRewardEntry"
        claimed:
          type: boolean
        isToday:
          type: boolean
    DailyRewardStatus:
      type: object
      required: [alreadyClaimed, currentStreak, nextStreakDay, reward, calendar]
      properties:
        alreadyClaimed:
          type: boolean
        currentStreak:
          type: integer
        nextStreakDay:
          type: integer
        reward:
          $ref: "#/components/schemas/DailyRewardEntry"
        calendar:
          type: array
          items:
            $ref: "#/components/schemas/DailyRewardCalendarDay"
        lastClaimedDate:
          type: ["string", "null"]
    DailyRewardClaimResult:
      type: object
      required: [success, streakDay, rewardType, rewardAmount, rewardLabel]
      properties:
        success:
          type: boolean
        streakDay:
          type: integer
        rewardType:
          type: string
        rewardAmount:
          type: integer
        rewardLabel:
          type: string
        isMilestone:
          type: boolean
...